+++ examples/jerni-store-with-dev/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="1.0.53"></a>
+## [1.0.53](https://github.com/tungv/jerni/compare/heq-store-with-dev@1.0.52...heq-store-with-dev@1.0.53) (2019-05-06)
+
+
+
+
+**Note:** Version bump only for package heq-store-with-dev
+
+<a name="1.0.52"></a>
+## [1.0.52](https://github.com/tungv/jerni/compare/heq-store-with-dev@1.0.51...heq-store-with-dev@1.0.52) (2019-05-02)
+
+
+
+
+**Note:** Version bump only for package heq-store-with-dev
+
+<a name="1.0.51"></a>
+## [1.0.51](https://github.com/tungv/jerni/compare/heq-store-with-dev@1.0.50...heq-store-with-dev@1.0.51) (2019-05-02)
+
+
+
+
+**Note:** Version bump only for package heq-store-with-dev
+
 <a name="1.0.50"></a>
 ## [1.0.50](https://github.com/tungv/jerni/compare/heq-store-with-dev@1.0.49...heq-store-with-dev@1.0.50) (2019-04-28)