+++ packages/heq-server-redis/CHANGELOG.md
@@ -3,6 +3,29 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="2.0.2"></a>
+## [2.0.2](https://github.com/tungv/events/compare/@heq/server-redis@2.0.1...@heq/server-redis@2.0.2) (2019-05-06)
+
+
+### Bug Fixes
+
+* **server-redis:** only apply index if fully indexed ([2b3641e](https://github.com/tungv/events/commit/2b3641e))
+
+
+
+
+<a name="2.0.1"></a>
+## [2.0.1](https://github.com/tungv/events/compare/@heq/server-redis@2.0.0...@heq/server-redis@2.0.1) (2019-05-02)
+
+
+### Bug Fixes
+
+* **server-redis:** avoid race condition  and missing event ([55038d7](https://github.com/tungv/events/commit/55038d7))
+* **server-redis:** clean up speculative queue ([13adfce](https://github.com/tungv/events/commit/13adfce))
+
+
+
+
 <a name="1.0.4"></a>
 ## [1.0.4](https://github.com/tungv/events/compare/@heq/server-redis@1.0.3...@heq/server-redis@1.0.4) (2018-07-20)